Update angular version to 1.8.2 in `setup.cfg` file.
This patch updates the angular version from 1.5.8->1.8.2 and
also replaces dashes with underscores in `setup.cfg` file.
The reason for replacing dashes with underscores is Setuptools
v54.1.0 introduces a warning that the use of dash-separated
options in 'setup.cfg' will not be supported in a future version [1].
Get ahead of the issue by replacing the dashes with underscores. Without
this, we see 'UserWarning' messages like the following on new enough
versions of setuptools:

  UserWarning: Usage of dash-separated 'description-file' will not be
  supported in future versions. Please use the underscore name
  'description_file' instead

[1] https://github.com/pypa/setuptools/commit/a2e9ae4cb

Change-Id: I4b64d5b8aa3841b3cac8d02819401400c247ff46